[UPD] loan_application module
